# mtd: asynchronous rcu quiesce for reducing latency